Move files to jwt/http/
