+++ packages/jwt/src/http/jwt-optional.hook.ts
